feat: unify db access and update docs
