+++ README.md
@@ -10,6 +10,7 @@
 
 Дополнительная информация о плане разработки представлена в файле `ROADMAP.md` (включая раздел «В разработке»), а история изменений ведётся в `CHANGELOG.md`.
 За инструкцией по настройке команд и вебхуков бота обратитесь к файлу `docs/telegram_bot_manual.md`. Документация API приведена в `docs/api_reference.md` и доступна через `/api-docs`.
+Карта запросов между API и базой описана в `docs/db_request_map.md`.
 
 ## Быстрый старт
