+++ docs/telegram_bot_manual.md
@@ -58,6 +58,16 @@ bot.sendMessage(chatId, 'Выберите действие', {
 
 ## Регистрация и администрирование
 Бот автоматически сохраняет пользователя при `/start`. Администраторы используют `/list_users` и `/add_user` для управления списком пользователей.
+Ниже приведены основные команды:
+- `/create_task` — создать задачу
+- `/assign_task` — назначить задачу пользователю
+- `/list_tasks` — мои задачи
+- `/list_all_tasks` — все задачи (только для админа)
+- `/update_task_status` — сменить статус задачи
+- `/upload_file` — отправить текстовый файл в R2
+- `/send_photo` — отправить фото по URL
+- `/edit_last` — редактировать сообщение
+- `/app` — открыть мини-приложение
 
 ## Диагностика
 Запрос `getMe` проверяет работоспособность токена:
